Implement non-traversable slopes and refactor
Note that while this is works, it is not perfect; when transitioning between the ground and a slope, entities become
"stuck" briefly, and we have to rely on the unstuck logic to free them.

This is a consequence of the way we handle multiple collisions occurring in a single frame. Since we process them
sequentially and the entity gets moved after each collision, subsequent collision results get skewed because we are
no longer simulating the exact movement that the entity has taken.